requirements.txt
streamlit==1.37.1
faster-whisper==1.0.3
sentence-transformers==3.0.1
chromadb==0.5.5
rank-bm25==0.2.2
pydub==0.25.1
numpy==1.26.4
scikit-learn==1.4.2
pandas==2.2.2
pyannote.audio==3.1.1
torch==2.3.1
groq==0.11.0
openai==1.40.6
tqdm==4.66.4
